requirements.txt
Django
pycryptodome